+++ src/renderer/src/views/x/AccountXView.vue
@@ -20,7 +20,7 @@ import SpeechBubble from '../shared_components/SpeechBubble.vue';
 import XProgressComponent from './XProgressComponent.vue';
 import XJobStatusComponent from './XJobStatusComponent.vue';
 
-import XWizardImportOrBuildPage from './XWizardImportOrBuildPage.vue';
+import XWizardDatabasePage from './XWizardDatabasePage.vue';
 import XWizardImportPage from './XWizardImportPage.vue';
 import XWizardImportDownloadPage from './XWizardImportDownloadPage.vue';
 import XWizardImportingPage from './XWizardImportingPage.vue';
@@ -455,8 +455,8 @@ onUnmounted(async () => {
         <div :class="{ 'hidden': model.showBrowser || model.state == State.RunJobs, 'wizard': true, 'ms-2': true }">
             <div class="wizard-container d-flex">
                 <div class="wizard-content flex-grow-1">
-                    <XWizardImportOrBuildPage v-if="model.state == State.WizardImportOrBuildDisplay"
-                        :model="unref(model)" @set-state="setState($event)" />
+                    <XWizardDatabasePage v-if="model.state == State.WizardDatabaseDisplay" :model="unref(model)"
+                        @set-state="setState($event)" />
 
                     <XWizardImportPage v-if="model.state == State.WizardImportStartDisplay" :model="unref(model)"
                         @set-state="setState($event)" />
