+++ src/renderer/src/views/x/XWizardArchiveOptionsPage.vue
@@ -23,7 +23,7 @@ const nextClicked = async () => {
 
 const backClicked = async () => {
     await saveSettings();
-    emit('setState', State.WizardImportOrBuild);
+    emit('setState', State.WizardDatabase);
 };
 
 // Settings
