+++ src/renderer/src/views/x/XWizardDatabasePage.vue
