+++ src/renderer/src/views/x/XWizardDeleteOptionsPage.vue
@@ -150,8 +150,8 @@ onMounted(async () => {
             </p>
         </div>
 
-        <XLastImportOrBuildComponent :account-i-d="model.account.id" :button-text="'Import or Build Database Again'"
-            :button-text-no-data="'Import or Build Database First'" :button-state="State.WizardImportOrBuild"
+        <XLastImportOrBuildComponent :account-i-d="model.account.id" :button-text="'Build Local Database Again'"
+            :button-text-no-data="'Build Local Database First'" :button-state="State.WizardDatabase"
             @set-state="emit('setState', $event)" />
 
         <form @submit.prevent>
