+++ src/renderer/src/views/x/XWizardSidebar.vue
@@ -87,7 +87,7 @@ onMounted(async () => {
     <div class="wizard-sidebar">
         <ul class="wizard-nav">
             <li>
-                <button class="btn btn-light" @click="emit('setState', State.WizardImportOrBuild)">
+                <button class="btn btn-light" @click="emit('setState', State.WizardDatabase)">
                     <i class="fa-solid fa-database" />
                     Local Database
                 </button>
